backend/app.py
import base64
import requests
import sqlite3
import os
from flask import Flask, request, jsonify
from flask_cors import CORS
import pickle
from datetime import datetime
from features import extract_features 
from dotenv import load_dotenv

# Load Env Vars
# Load Env Vars
load_dotenv()
from local_brain import LocalBrain

# Initialize Local Engine
brain = LocalBrain()

app = Flask(__name__)
CORS(app, resources={r"/*": {"origins": "*"}})

# Configuration
VT_API_KEY = os.getenv("VT_API_KEY")
DB_FILE = "phishing.db"

# --- DATABASE SETUP ---
def init_db():
    with sqlite3.connect(DB_FILE) as conn:
        cursor = conn.cursor()
        cursor.execute('''
            CREATE TABLE IF NOT EXISTS scan_history (
                id INTEGER PRIMARY KEY AUTOINCREMENT,
                url TEXT NOT NULL,
                status TEXT,
                score INTEGER,
                timestamp DATETIME DEFAULT CURRENT_TIMESTAMP,
                source TEXT
            )
        ''')
        conn.commit()

init_db()

# Load Local AI
try:
    with open('classifier.pkl', 'rb') as f:
        model = pickle.load(f)
except FileNotFoundError:
    model = None
    print("⚠️ Warning: classifier.pkl not found. Local AI disabled.")

def save_scan_result(url, status, score, source):
    """Save scan to SQLite database"""
    try:
        with sqlite3.connect(DB_FILE) as conn:
            cursor = conn.cursor()
            cursor.execute("INSERT INTO scan_history (url, status, score, source) VALUES (?, ?, ?, ?)", 
                           (url, status, score, source))
            conn.commit()
    except Exception as e:
        print(f"❌ DB Error: {e}")

def get_virustotal_report(url):
    print(f"--- 🚀 STARTING VIRUSTOTAL LOOKUP FOR: {url} ---")
    
    if not VT_API_KEY:
        print("❌ ERROR: VT_API_KEY missing in .env")
        return None

    try:
        url_id = base64.urlsafe_b64encode(url.encode()).decode().strip("=")
        endpoint = f"https://www.virustotal.com/api/v3/urls/{url_id}"
        headers = {"x-apikey": VT_API_KEY}

        print(f"📡 Sending request to: {endpoint}")
        response = requests.get(endpoint, headers=headers)
        
        if response.status_code == 200:
             print("✅ Data received successfully!")
             data = response.json()
             attr = data['data']['attributes']
             stats = attr['last_analysis_stats']
             
             # Vendor Logic
             vendor_list = []
             for name, result in attr.get('last_analysis_results', {}).items():
                 vendor_list.append({
                     "name": name,
                     "status": result['category'],
                     "result": result['result']
                 })
 
             return {
                 "source": "SHIELD.AI Global Threat Grid",
                 "malicious": stats['malicious'],
                 "suspicious": stats['suspicious'],
                 "harmless": stats['harmless'],
                 "reputation": attr.get('reputation', 0),
                 "scan_date": attr.get('last_analysis_date', int(datetime.now().timestamp())),
                 "vendors": vendor_list,
                 "categories": attr.get('categories', {}),
                 "final_url": attr.get('final_url', url),
                 "http_code": attr.get('last_http_response_code', "N/A"),
                 "serving_ip": attr.get('last_http_response_ip_address', "Hidden/Cloud"),
                 "headers": attr.get('last_http_response_headers', {}),
                 "meta_tags": attr.get('html_meta', {}),
                 "history": {
                     "first": datetime.utcfromtimestamp(attr.get('first_submission_date', 0)).strftime('%Y-%m-%d %H:%M:%S') if attr.get('first_submission_date') else "Unknown",
                     "last": datetime.utcfromtimestamp(attr.get('last_submission_date', 0)).strftime('%Y-%m-%d %H:%M:%S') if attr.get('last_submission_date') else "Unknown"
                 }
             }
         
        elif response.status_code == 404:
             print("⚠️ URL not found in Cloud database.")
             return None 
        elif response.status_code == 401:
             print("❌ ERROR: Invalid API Key.")
             return None
        else:
             print(f"❌ ERROR: Connection failed. Status: {response.status_code}")
             return None
 
    except requests.exceptions.ConnectionError:
        print("🌐⚠️ INTERNET DISCONNECTED: Connection to VirusTotal failed.")
        print("🔄 AUTOMATIC SWITCH: Engaging Local Neural Engine & Blacklist...")
        return None
    except Exception as e:
        print(f"❌ CRITICAL ERROR: {e}")
        return None

@app.route('/health', methods=['GET'])
def health_check():
    return jsonify({"status": "online"}), 200

@app.route('/history', methods=['GET'])
def get_history():
    """Retrieve last 50 scans from DB"""
    try:
        with sqlite3.connect(DB_FILE) as conn:
            cursor = conn.cursor()
            cursor.execute("SELECT url, status, score, timestamp, source FROM scan_history ORDER BY id DESC LIMIT 50")
            rows = cursor.fetchall()
            
            history_data = []
            for row in rows:
                history_data.append({
                    "url": row[0],
                    "status": row[1],
                    "score": row[2],
                    "time": row[3],
                    "vt_stats": True if row[4] == "SHIELD.AI Global Threat Grid" else False 
                })
            return jsonify(history_data), 200
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route('/history', methods=['DELETE'])
def clear_history():
    """Clear all scan history"""
    try:
        with sqlite3.connect(DB_FILE) as conn:
            cursor = conn.cursor()
            cursor.execute("DELETE FROM scan_history")
            conn.commit()
            return jsonify({"status": "History Cleared"}), 200
    except Exception as e:
        return jsonify({"error": str(e)}), 500

@app.route('/predict', methods=['POST'])
def predict():
    data = request.get_json()
    url = data.get('url')
    mode = data.get('mode', 'online') # Default to online
    
    if not url: return jsonify({"error": "No URL"}), 400

    # --- 🛡️ DEMO TRAINING DATA (DYNAMIC LOADING) ---
    DEMO_PHISHING_LINKS = []
    try:
        dataset_path = os.path.join(os.path.dirname(__file__), 'datasets', 'phishing_training.txt')
        if os.path.exists(dataset_path):
            with open(dataset_path, 'r') as f:
                DEMO_PHISHING_LINKS = [line.strip() for line in f.readlines() if line.strip()]
        else:
             print(f"⚠️ Warning: Dataset not found at {dataset_path}. Using fallback demo list.")
             DEMO_PHISHING_LINKS = [
                "http://paypal-secure-login-update.com",
                "http://google-drive-secure-share.com"
             ]
    except Exception as e:
        print(f"❌ Error loading training data: {e}")

    if url in DEMO_PHISHING_LINKS:
        print(f"🚨 PHISHING LINK MATCHED: {url} found in Local Blacklist.")
        
        # Simulate Rich Forensic Data for Demo
        mock_vendors = [
            {"name": "SHIELD.AI Heuristics", "status": "malicious", "result": "phishing"},
            {"name": "Local Neural Core", "status": "malicious", "result": "malware"},
            {"name": "Pattern Matcher", "status": "suspicious", "result": "suspicious"},
            {"name": "Domain Reputation", "status": "malicious", "result": "poor_reputation"}
        ]
        
        mock_stats = {
            "source": "SHIELD.AI Local Cortex (Enhanced)",
            "malicious": 85,
            "suspicious": 12,
            "harmless": 0,
            "reputation": -90,
            "scan_date": int(datetime.now().timestamp()),
            "vendors": mock_vendors,
            "final_url": url,
            "http_code": 200,
            "serving_ip": "127.0.0.1 (Simulated)",
            "headers": {"server": "Apache/2.4.52 (Debian)", "content-type": "text/html; charset=UTF-8"},
            "meta_tags": {"viewport": "width=device-width", "author": "admin"},
            "categories": {"Forcepoint ThreatSeeker": "Phishing", "Sophos": "Malware"},
            "history": {"first": "2024-01-01 00:00:00", "last": datetime.now().strftime('%Y-%m-%d %H:%M:%S')}
        }

        result = {
            "status": "Phishing Detected",
            "score": 99,
            "verdict": "Phishing Detected",
            "details": "⚠️ CRITICAL THREAT: Identified in Local Blacklist Database.\nHigh-confidence match with known phishing signatures.",
            "reasons": ["Match found in Training Dataset", "High-Confidence Threat Signature", "Domain Blacklisted", "Heuristic Analysis Failed"],
            "geo": {"origin": "Local Database", "threat_type": "Internal Blacklist"},
            "vt_stats": mock_stats # Inject Mock Data for Charts
        }
        status_save = "Phishing Detected"
        save_scan_result(url, status_save, 99, "SHIELD.AI Internal Blacklist")
        return jsonify(result)

    # MODE: OFFLINE (Local Only)
    if mode == 'offline':
        print("⚡ MODE: OFFLINE. Using Local Cortex Only.")
        result = brain.analyze(url)
        result['status'] = result['verdict'] # Normalize for Frontend
        save_scan_result(url, result['verdict'], result['score'], "Shield.AI Local Cortex")
        return jsonify(result)

    # MODE: ONLINE (Hybrid)
    # 1. VIRUSTOTAL CHECK
    vt_data = get_virustotal_report(url)
    
    if vt_data:
        # --- ✅ NEW SMART LOGIC (Matches Online Version) ---
        malicious = vt_data['malicious']
        suspicious = vt_data['suspicious']

        # RULE 1: If even 1 vendor says "Malicious", flag it.
        if malicious > 0:
            status = "Phishing Detected"
            score = 85 + (malicious * 5) # Scale up with more detections (Max 100)
            if score > 100: score = 100

        # RULE 2: Ignore just 1 "Suspicious" flag (False Positive protection).
        elif suspicious > 1:
            status = "Suspicious"
            score = 65
            
        # RULE 3: Otherwise, it is Safe.
        else:
            status = "Clean"
            score = 10 # Low risk score for safe sites

        # SAVE TO DB
        save_scan_result(url, status, score, "SHIELD.AI Global Threat Grid")

        return jsonify({
            "status": status,
            "score": score,
            "details": f"Verified by Global Grid. {vt_data['harmless']} security nodes confirm safety.",
            "reasons": [f"Flagged by {malicious} vendors"] if malicious > 0 else ["Clean Reputation"],
            "geo": {"origin": "Global Cloud", "threat_type": "External Intel"},
            "vt_stats": vt_data
        })
        
    else:
        # Fallback if VT fails or API key invalid
        print("⚡ OFFLINE FALLBACK: Using Local Neural Engine.")
        result = brain.analyze(url)
        result['status'] = result['verdict'] # Normalize for Frontend
        save_scan_result(url, result['verdict'], result['score'], "Shield.AI Local Cortex (Fallback)")
        return jsonify(result)

if __name__ == '__main__':
    # Localhost Port 5000
    app.run(debug=True, port=5000, host='0.0.0.0')